requirements.txt
asgiref==3.7.2
channels==4.0.0
click==8.1.6
colorama==0.4.6
Django==4.2.3
djangorestframework==3.14.0
h11==0.14.0
psycopg2==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
uvicorn==0.23.1
